chore(leaderboard): Break down leaderboard table
