--- apps/leaderboard/src/features/leaderboard/components/BadgesCell/index.ts
@@ -1,2 +1,0 @@
-export * from './badges';
-export * from './BadgesCell';
