+++ apps/leaderboard/src/features/leaderboard/types/index.ts
@@ -1,4 +1,4 @@
-import { BadgeEnum } from '../components';
+import { BadgeEnum } from '../constants';
 
 export interface PointsBreakdown {
   mainnet: bigint;
